+++ packages/react-ai-sdk/src/ui/getVercelAIMessages.tsx
@@ -1,9 +1,9 @@
 import {
-  getExternalStoreMessage,
+  getExternalStoreMessages,
   type ThreadMessage,
 } from "@assistant-ui/react";
 import type { Message } from "@ai-sdk/ui-utils";
 
 export const getVercelAIMessages = (message: ThreadMessage) => {
-  return getExternalStoreMessage(message) as Message[];
+  return getExternalStoreMessages(message) as Message[];
 };