+++ packages/create-assistant-ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-assistant-ui
 
+## 0.0.21
+
+### Patch Changes
+
+- 98a680e: chore: update deps
+
 ## 0.0.20
 
 ### Patch Changes